requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-environ==0.11.2
django-redis==5.3.0
idna==3.4
redis==5.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
